refactore index.js, App.js
