--- client/src/containers/App.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import PatientBoard from './PatientBoard';
-import TranslationsSuspense from '../components/TranslationsSuspense';
-
-const App = () => (
-  <TranslationsSuspense>
-    <PatientBoard />
-  </TranslationsSuspense>
-);
-export default App;
